+++ internal/service/diff/commit_test.go
@@ -6,11 +6,9 @@ import (
 	"io"
 	"net"
 	"testing"
-	"time"
 
 	pb "gitlab.com/gitlab-org/gitaly-proto/go"
 	"gitlab.com/gitlab-org/gitaly/internal/diff"
-	"gitlab.com/gitlab-org/gitaly/internal/service/renameadapter"
 	"gitlab.com/gitlab-org/gitaly/internal/testhelper"
 
 	"github.com/stretchr/testify/require"
@@ -882,27 +880,12 @@ func runDiffServer(t *testing.T) *grpc.Server {
 		t.Fatal(err)
 	}
 
-	pb.RegisterDiffServer(server, renameadapter.NewDiffAdapter(NewServer()))
+	pb.RegisterDiffServiceServer(server, NewServer())
 	reflection.Register(server)
 
 	go server.Serve(listener)
 
 	return server
-}
-
-func newDiffClient(t *testing.T) (pb.DiffClient, *grpc.ClientConn) {
-	connOpts := []grpc.DialOption{
-		grpc.WithInsecure(),
-		grpc.WithDialer(func(addr string, _ time.Duration) (net.Conn, error) {
-			return net.Dial("unix", addr)
-		}),
-	}
-	conn, err := grpc.Dial(serverSocketPath, connOpts...)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	return pb.NewDiffClient(conn), conn
 }
 
 func drainCommitDiffResponse(c pb.Diff_CommitDiffClient) error {